GPIO: GPIO Read function returns 8 bits

+++ src/hw/gpio/gpio.h
@@ -125,7 +125,7 @@ void hwGpioSetOutput( Uint32 uiNumber);
 void hwGpioClearOutput( Uint32 uiNumber);
 Uint32 hwGpioReadInput( Uint32 uiNumber );
 void hwGpioWriteDataBus ( Uint8 uchValue );
-Uint32 hwGpioReadDataBus( void );
+Uint8 hwGpioReadDataBus( void );
 void hwGpioEnableGlobalInterrupt( void );
 void hwGpioSetDataBus(Uint8 uchValue);
 void hwGpioClearDataBus(Uint8 uchValue);
